+++ tests/testthat/test-age-categorical.R
@@ -1,5 +1,5 @@
 test_that("age_cat_fun performs as expected below lower bound", {
-  expect_equal(age_cat_fun(11), NA)
+  expect_equal(age_cat_fun(11), "NA(b)")
 })
 
 test_that("age_cat_fun performs as expected within bounds", {