+++ NEWS.md
@@ -1,4 +1,40 @@
-# cchsflow 1.7.1 (Latest build)
+# cchsflow 1.8.0
+2021-01-22
+
+## Features
+- New function `merge_rec_data()` that merges and labels transformed CCHS
+datasets.
+
+## Minor improvements
+- Improved support for `tagged_na` in derived variables. Additional
+documentation added for creating derived variables in the [tagged_na](https://big-life-lab.github.io/cchsflow/articles/tagged_na_usage.html)
+article.
+- Updates to labels and variable classification on `variables.csv` and
+`variable_details.csv`.
+
+## New variables
+- [**COPD_Emph_der:**](https://big-life-lab.github.io/reference/COPD_Emph_der_fun1.html)
+derived variable that determines if a respondent has either COPD or Emphysema.
+- [**ADL_score_5:**](https://big-life-lab.github.io/reference/adl_score_5_fun.html)
+derived variable that scores the number of daily tasks that a respondent needs
+help with.
+- [**diet_score:**](https://big-life-lab.github.io/reference/diet_score_fun.html)
+derived variable that scores respondents based on daily consumption of fruit,
+vegetables, and fruit juice.
+- [**smoke_simple:**](https://big-life-lab.github.io/reference/smoke_simple_fun.html)
+derived variable that identifies respondents smoking status.
+- [**time_quit_smoking:**](https://big-life-lab.github.io/reference/time_quit_smoking_fun.html)
+derived variable that estimates approximate time since respondent has quit
+smoking if they are a former smoker.
+- [**LBFA_31A/LBFA_31A_a/LBFA_31A_b:**](https://big-life-lab.github.io/reference/LBFA_31A.html)
+Occupation group. 3 different variables with differing categories were created
+to harmonize between CCHS cycles.
+- **CCC_041:** Fibromyalgia.
+- **CCC_061:** Back problems.
+- **CCC_081:** Migraine headaches.
+- **CCC_290:** Anxiety disorder.
+
+# cchsflow 1.7.1
 2020-09-16
 
 ## Minor improvements